+++ src/Exception/FileNotExistsException.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * This file is part of php-tools.
  *
@@ -21,13 +22,38 @@ use Exception;
 class FileNotExistsException extends Exception
 {
     /**
-     * Constructor
-     * @param string $message The string of the error message
-     * @param int $code The code of the error
-     * @param \Throwable $previous the previous exception
+     * @var string|null
      */
-    public function __construct($message = 'File or directory does not exist', $code = 0, $previous = null)
+    protected $path;
+
+    /**
+     * Constructor
+     * @param string|null $message The string of the error message
+     * @param int $code The code of the error
+     * @param \Throwable|null $previous the previous exception
+     * @param string|null $path Path of file that do not exist
+     * @uses $path
+     */
+    public function __construct($message = null, $code = 0, \Throwable $previous = null, $path = null)
     {
+        if (!$message) {
+            $message = 'File or directory does not exist';
+            if ($path) {
+                $message = sprintf('File or directory `%s` does not exist', $path);
+            }
+        }
         parent::__construct($message, $code, $previous);
+        $this->path = $path;
+    }
+
+    /**
+     * Gets the path of file that do not exist
+     * @return string|null
+     * @since 1.2.10
+     * @uses $path
+     */
+    public function getFilePath()
+    {
+        return $this->path;
     }
 }