+++ src/Exception/NotDirectoryException.php
@@ -22,13 +22,38 @@ use Exception;
 class NotDirectoryException extends Exception
 {
     /**
-     * Constructor
-     * @param string $message The string of the error message
-     * @param int $code The code of the error
-     * @param \Throwable $previous the previous exception
+     * @var string|null
      */
-    public function __construct($message = 'Filename is not a directory', $code = 0, $previous = null)
+    protected $path;
+
+    /**
+     * Constructor
+     * @param string|null $message The string of the error message
+     * @param int $code The code of the error
+     * @param \Throwable|null $previous the previous exception
+     * @param string|null $path Path of filename that is not a directory
+     * @uses $path
+     */
+    public function __construct($message = null, $code = 0, \Throwable $previous = null, $path = null)
     {
+        if (!$message) {
+            $message = 'Filename is not a directory';
+            if ($path) {
+                $message = sprintf('Filename `%s` is not a directory', $path);
+            }
+        }
         parent::__construct($message, $code, $previous);
+        $this->path = $path;
+    }
+
+    /**
+     * Gets the path of filename that is not a directory
+     * @return string|null
+     * @since 1.2.10
+     * @uses $path
+     */
+    public function getFilePath()
+    {
+        return $this->path;
     }
 }